requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.15
aiosignal==1.3.2
alembic==1.15.2
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1732732445543/work
annotated-types==0.7.0
anyio==4.9.0
archspec @ file:///croot/archspec_1709217642129/work
arxiv==2.1.3
asgiref==3.8.1
asttokens==3.0.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
boltons @ file:///croot/boltons_1737061692168/work
Brotli==1.1.0
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.2
cairocffi==1.7.1
CairoSVG==2.8.2
certifi @ file:///croot/certifi_1738623731865/work/certifi
cffi @ file:///croot/cffi_1736182485317/work
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
cloudpickle==3.1.1
coloredlogs==15.0.1
comm==0.2.2
conda @ file:///croot/conda_1738168370816/work
conda-anaconda-telemetry @ file:///croot/conda-anaconda-telemetry_1736524596117/work
conda-anaconda-tos @ file:///croot/conda-anaconda-tos_1739298990519/work
conda-content-trust @ file:///croot/conda-content-trust_1714483159009/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1737733694612/work/src
conda-package-handling @ file:///croot/conda-package-handling_1731369017509/work
conda_package_streaming @ file:///croot/conda-package-streaming_1731366181659/work
contourpy==1.3.2
cryptography @ file:///croot/cryptography_1732129644106/work
cssselect2==0.8.0
cycler==0.12.1
databricks-sdk==0.50.0
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
distro @ file:///croot/distro_1714488253808/work
docker==7.1.0
duckduckgo_search==5.3.1b1
durationpy==0.9
executing==2.2.0
faiss-cpu==1.10.0
fastapi==0.115.12
feedparser==6.0.11
filelock==3.18.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.57.0
frozendict @ file:///croot/frozendict_1713194832637/work
frozenlist==1.5.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
googleapis-common-protos==1.69.2
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.1
groq==0.20.0
grpcio==1.71.0
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.1
humanfriendly==10.0
hyperframe==6.1.0
idna @ file:///croot/idna_1714398848350/work
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
kubernetes==32.0.1
langchain==0.3.22
langchain-chroma==0.2.2
langchain-community==0.3.20
langchain-core==0.3.49
langchain-groq==0.3.2
langchain-huggingface==0.1.2
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langsmith==0.3.21
libmambapy @ file:///croot/mamba-split_1734469461757/work/libmambapy
Mako==1.3.10
Markdown==3.8
markdown-it-py @ file:///work/perseverance-python-buildout/croot/markdown-it-py_1698846045803/work
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl @ file:///work/perseverance-python-buildout/croot/mdurl_1698845653285/work
menuinst @ file:///croot/menuinst_1738943416351/work
mlflow==2.22.0
mlflow-skinny==2.22.0
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.3.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nodejs==0.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
onnxruntime==1.21.0
openai==1.70.0
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
optional-django==0.1.0
orjson==3.10.16
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==12.1.0
platformdirs==4.3.7
pluggy @ file:///croot/pluggy_1733169602837/work
posthog==3.23.0
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycosat @ file:///croot/pycosat_1736868416091/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.11.1
pydantic-settings==2.8.1
pydantic_core==2.33.0
pydub==0.25.1
Pygments @ file:///work/perseverance-python-buildout/croot/pygments_1698846270603/work
PyMuPDF==1.25.5
pypdf==5.4.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dotenv==1.1.0
pyzmq==26.3.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich @ file:///croot/rich_1732638981168/work
rsa==4.9
ruamel.yaml @ file:///croot/ruamel.yaml_1727980156079/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1727769819935/work
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.1
setuptools==75.8.0
sgmllib3k==1.0.0
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
SQLAlchemy==2.0.40
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.1
sympy==1.14.0
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.5.1
tokenizers==0.21.1
torch==2.7.0
torchaudio==2.7.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.3
triton==3.3.0
truststore @ file:///croot/truststore_1736550121485/work
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3 @ file:///croot/urllib3_1737133630106/work
uvicorn==0.34.0
uvloop==0.21.0
validators==0.28.1
watchfiles==1.0.4
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wikipedia==1.4.0
wrapt==1.17.2
yarl==1.18.3
yt-dlp==2025.5.22
zipp==3.21.0
zstandard==0.23.0
